api/requirements.txt
Flask==3.1.0
Flask-Cors==Flask-Cors
SQLAlchemy==SQLAlchemy
cryptography==44.0.0
psycopg==3.2.3
google-auth==2.37.0